pyproject.toml
[project]
name = "democratic-machine-learning"
version = "0.1.0"
description = "AI-augmented democratic governance framework with LLM-driven policy analysis"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "numpy>=1.26",
    "playwright>=1.40",
    "prompt-toolkit>=3.0.52",
    "pyyaml>=6.0",
    "requests>=2.32",
    "rich>=13.0",
]

[dependency-groups]
dev = [
    "mypy>=1.0",
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "typeguard>=4.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
